query breakdown tests now all pass
